requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
GDAL==3.7.3
django-extensions==3.2.3
python-dotenv==1.0.0
Pillow==10.1.0
gunicorn==21.2.0
whitenoise==6.6.0
